Add the CDC-ACM class again.
Let CubeMX generate the entire framework of the USB-Device core, and regristration of class and interface.
In the next step I will tweak the generated code to replace the CDC-ACM class with CDC-RNDIS class.

This approach should allow to use this code base and CubeMX project to migrate it to other MCU types, and keeping up-to-date with
new Cube library version by just re-generating the code.