Core/Src/usbd_cdc_rndis_if.c: Make it work!!!
Major error was that I need to leave enough space at the start of the
send buffer for the RNDIS header (sizeof(USBD_CDC_RNDIS_PacketMsgTypeDef)),
and increase the length to be sent accordingly.
Before no sent package went through properly because the RNDIS header
overlapped with the ethernet content.

Activate and initialize autoip.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>