Switch the USB class from RNDIS to CDC-ECM.
Mostly mechanical task: The structures and functions, particularly in Core/Src/usbd_cdc_rndis_if.c were almost identical.
Migrating my additional code in this module was trivial.
Otherwise replace RNDIS with ECM, and that was it.

The only difference of notice is that ECM obviously does not add a kind of "ECM header" at the start of a package like RNDIS does.