USB_DEVICE/Target/usbd_conf.h: Define USBD_SUPPORT_USER_STRING_DESC=1U
By default USBD_SUPPORT_USER_STRING_DESC is defined 0 in usbd_def.h.
The user string for ECM is the Ethernet address string which is
essential part of the interface initialization. Without it Linux could
not initialize the interface.

Not the interface starts, and I can ping the STM controller via USB.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>